GSYE-416-refactor: Move area classes into separate modules
